Fix library path
We need to set `R_ENVIRON_USER` to an existing file,
otherwise R does not use it. We also need to set R_LIBS
to empty, because otherwise R will pick it up.

Closes #160.

Note that this is also fixed in callr, once
https://github.com/r-lib/callr/pull/84 is merged.
The callr fix is not needed for this revdepcheck fix.