use `force=TRUE` during revdep_install
now that I'm thinking more about it, the issue might come from somewhere else  (maybe `with_libpaths()`)... but in my case, `install_packages()` would not install the "new" version of the package because it detected that the `SHA1` hadn't changed, leaving `revdep/library/{pkg}/new/{pkg}` empty.



The checks continued with the final result making it seem that all tests failed (actually due to the absence of the "new" version of the package in the library). Would adding checks to ensure that both the `new` and the `old` libraries are not empty be pertinent as well?