Cope with lack of DESCRIPTION
I had an authentic example of this, which prevents you from using `cloud_summary()` or `cloud_report()` at all. It was around the time of the AWS outage, so perhaps it's a rare event, but apparently it can happen.

Here's what failure looks like for both `cloud_summary()` and `cloud_report()`:

> cloud_summary()
ℹ Syncing results to revdep/cloud.noindex/d61b319f-1c0e-42df-926e-89cc5c0c140b
ℹ Comparing results
Error in idesc_create_file(self, private, file) : 
  is_existing_file(file) is not TRUE